refactor: 유저 프로필 페이지 컴포넌트 생성
